% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\title{Compute the students' score of spheredata package}
\usage{
binary(raw, key)
}
\arguments{
\item{raw}{a dataframe of raw response data}

\item{key}{a dataframe of answer key}
}
\value{
a dataframe of dichotomous format of students' response data
}
\description{
Compute the students' score as binary/ dichotomous value. The calculation
method is based on a classical test theory prespective.
}
\examples{
# Import the FCI score and key data
library(spheredata)
data("FCI")
data("FCIkey")

# Processing the Force Concept Inventory (FCI) data as dichotomous
binary(FCI, FCIkey)
}
