\name{CLASScontentvalidity}
\alias{CLASScontentvalidity}
\docType{data}
\title{The Colorado Learning Attitudes about Science Survey (CLASS) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the CLASS.
}
\usage{data(CLASScontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 36 CLASS items (after preprocessed using the scoring rule).
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{CLASS1}}{A numeric of expert rating to the CLASS item 1.}
    \item{\code{CLASS2}}{A numeric of expert rating to the CLASS item 2.}
    \item{\code{CLASS3}}{A numeric of expert rating to the CLASS item 3.}
    \item{\code{CLASS5}}{A numeric of expert rating to the CLASS item 5.}
    \item{\code{CLASS6}}{A numeric of expert rating to the CLASS item 6.}
    \item{\code{CLASS8}}{A numeric of expert rating to the CLASS item 8.}
    \item{\code{CLASS10}}{A numeric of expert rating to the CLASS item 10.}
    \item{\code{CLASS11}}{A numeric of expert rating to the CLASS item 11.}
    \item{\code{CLASS12}}{A numeric of expert rating to the CLASS item 12.}
    \item{\code{CLASS13}}{A numeric of expert rating to the CLASS item 13.}
    \item{\code{CLASS14}}{A numeric of expert rating to the CLASS item 14.}
    \item{\code{CLASS15}}{A numeric of expert rating to the CLASS item 15.}
    \item{\code{CLASS16}}{A numeric of expert rating to the CLASS item 16.}
    \item{\code{CLASS17}}{A numeric of expert rating to the CLASS item 17.}
    \item{\code{CLASS18}}{A numeric of expert rating to the CLASS item 18.}
    \item{\code{CLASS19}}{A numeric of expert rating to the CLASS item 19.}
    \item{\code{CLASS20}}{A numeric of expert rating to the CLASS item 20.}
    \item{\code{CLASS21}}{A numeric of expert rating to the CLASS item 21.}
    \item{\code{CLASS22}}{A numeric of expert rating to the CLASS item 22.}
    \item{\code{CLASS23}}{A numeric of expert rating to the CLASS item 23.}
    \item{\code{CLASS24}}{A numeric of expert rating to the CLASS item 24.}
    \item{\code{CLASS25}}{A numeric of expert rating to the CLASS item 25.}
    \item{\code{CLASS26}}{A numeric of expert rating to the CLASS item 26.}
    \item{\code{CLASS27}}{A numeric of expert rating to the CLASS item 27.}
    \item{\code{CLASS28}}{A numeric of expert rating to the CLASS item 28.}
    \item{\code{CLASS29}}{A numeric of expert rating to the CLASS item 29.}
    \item{\code{CLASS30}}{A numeric of expert rating to the CLASS item 30.}
    \item{\code{CLASS32}}{A numeric of expert rating to the CLASS item 32.}
    \item{\code{CLASS34}}{A numeric of expert rating to the CLASS item 34.}
    \item{\code{CLASS35}}{A numeric of expert rating to the CLASS item 35.}
    \item{\code{CLASS36}}{A numeric of expert rating to the CLASS item 36.}
    \item{\code{CLASS37}}{A numeric of expert rating to the CLASS item 37.}
    \item{\code{CLASS38}}{A numeric of expert rating to the CLASS item 38.}
    \item{\code{CLASS39}}{A numeric of expert rating to the CLASS item 39.}
    \item{\code{CLASS40}}{A numeric of expert rating to the CLASS item 40.}
    \item{\code{CLASS42}}{A numeric of expert rating to the CLASS item 42.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("CLASScontentvalidity")
# Open the data
CLASScontentvalidity
}
