\name{FCIcontentvalidity}
\alias{FCIcontentvalidity}
\docType{data}
\title{The Force Concept Inventory (FCI) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the FCI.
}
\usage{data(FCIcontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 30 FCI items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{FCI1}}{A numeric value of expert rating to the FCI item number 1.}
    \item{\code{FCI2}}{A numeric value of expert rating to the FCI item number 2.}
    \item{\code{FCI3}}{A numeric value of expert rating to the FCI item number 3.}
    \item{\code{FCI4}}{A numeric value of expert rating to the FCI item number 4.}
    \item{\code{FCI5}}{A numeric value of expert rating to the FCI item number 5.}
    \item{\code{FCI6}}{A numeric value of expert rating to the FCI item number 6.}
    \item{\code{FCI7}}{A numeric value of expert rating to the FCI item number 7.}
    \item{\code{FCI8}}{A numeric value of expert rating to the FCI item number 8.}
    \item{\code{FCI9}}{A numeric value of expert rating to the FCI item number 9.}
    \item{\code{FCI10}}{A numeric value of expert rating to the FCI item number 10.}
    \item{\code{FCI11}}{A numeric value of expert rating to the FCI item number 11.}
    \item{\code{FCI12}}{A numeric value of expert rating to the FCI item number 12.}
    \item{\code{FCI13}}{A numeric value of expert rating to the FCI item number 13.}
    \item{\code{FCI14}}{A numeric value of expert rating to the FCI item number 14.}
    \item{\code{FCI15}}{A numeric value of expert rating to the FCI item number 15.}
    \item{\code{FCI16}}{A numeric value of expert rating to the FCI item number 16.}
    \item{\code{FCI17}}{A numeric value of expert rating to the FCI item number 17.}
    \item{\code{FCI18}}{A numeric value of expert rating to the FCI item number 18.}
    \item{\code{FCI19}}{A numeric value of expert rating to the FCI item number 19.}
    \item{\code{FCI20}}{A numeric value of expert rating to the FCI item number 20.}
    \item{\code{FCI21}}{A numeric value of expert rating to the FCI item number 21.}
    \item{\code{FCI22}}{A numeric value of expert rating to the FCI item number 22.}
    \item{\code{FCI23}}{A numeric value of expert rating to the FCI item number 23.}
    \item{\code{FCI24}}{A numeric value of expert rating to the FCI item number 24.}
    \item{\code{FCI25}}{A numeric value of expert rating to the FCI item number 25.}
    \item{\code{FCI26}}{A numeric value of expert rating to the FCI item number 26.}
    \item{\code{FCI27}}{A numeric value of expert rating to the FCI item number 27.}
    \item{\code{FCI28}}{A numeric value of expert rating to the FCI item number 28.}
    \item{\code{FCI29}}{A numeric value of expert rating to the FCI item number 29.}
    \item{\code{FCI30}}{A numeric value of expert rating to the FCI item number 30.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("FCIcontentvalidity")
# Open the data
FCIcontentvalidity
}
