\name{FCIkey}
\alias{FCIkey}
\docType{data}
\title{The Force Concept Inventory (FCI) key dataset}
\description{
The answers key to analyze the students' obtained score on the FCI.
}
\usage{data(FCIkey)}
\format{
  A data frame of a key record on the following 30 FCI items.
  \describe{
    \item{\code{FCI1}}{A key of the FCI item number 1.}
    \item{\code{FCI2}}{A key of the FCI item number 2.}
    \item{\code{FCI3}}{A key of the FCI item number 3.}
    \item{\code{FCI4}}{A key of the FCI item number 4.}
    \item{\code{FCI5}}{A key of the FCI item number 5.}
    \item{\code{FCI6}}{A key of the FCI item number 6.}
    \item{\code{FCI7}}{A key of the FCI item number 7.}
    \item{\code{FCI8}}{A key of the FCI item number 8.}
    \item{\code{FCI9}}{A key of the FCI item number 9.}
    \item{\code{FCI10}}{A key of the FCI item number 10.}
    \item{\code{FCI11}}{A key of the FCI item number 11.}
    \item{\code{FCI12}}{A key of the FCI item number 12.}
    \item{\code{FCI13}}{A key of the FCI item number 13.}
    \item{\code{FCI14}}{A key of the FCI item number 14.}
    \item{\code{FCI15}}{A key of the FCI item number 15.}
    \item{\code{FCI16}}{A key of the FCI item number 16.}
    \item{\code{FCI17}}{A key of the FCI item number 17.}
    \item{\code{FCI18}}{A key of the FCI item number 18.}
    \item{\code{FCI19}}{A key of the FCI item number 19.}
    \item{\code{FCI20}}{A key of the FCI item number 20.}
    \item{\code{FCI21}}{A key of the FCI item number 21.}
    \item{\code{FCI22}}{A key of the FCI item number 22.}
    \item{\code{FCI23}}{A key of the FCI item number 23.}
    \item{\code{FCI24}}{A key of the FCI item number 24.}
    \item{\code{FCI25}}{A key of the FCI item number 25.}
    \item{\code{FCI26}}{A key of the FCI item number 26.}
    \item{\code{FCI27}}{A key of the FCI item number 27.}
    \item{\code{FCI28}}{A key of the FCI item number 28.}
    \item{\code{FCI29}}{A key of the FCI item number 29.}
    \item{\code{FCI30}}{A key of the FCI item number 30.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Hestenes, D., Wells, M. & Swackhamer, G. Force concept inventory. \emph{Phys Teach 30}, 141–158 (1992).
}
\examples{
data(FCIkey)
}
\keyword{datasets}
