\name{FMCEcontentvalidity}
\alias{FMCEcontentvalidity}
\docType{data}
\title{The Force and Motion Conceptual Evaluation (FMCE) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the FMCE.
}
\usage{data(FMCEcontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 47 FMCE items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{FMCE1}}{A numeric value of expert rating to the FMCE item number 1.}
    \item{\code{FMCE2}}{A numeric value of expert rating to the FMCE item number 2.}
    \item{\code{FMCE3}}{A numeric value of expert rating to the FMCE item number 3.}
    \item{\code{FMCE4}}{A numeric value of expert rating to the FMCE item number 4.}
    \item{\code{FMCE5}}{A numeric value of expert rating to the FMCE item number 5.}
    \item{\code{FMCE6}}{A numeric value of expert rating to the FMCE item number 6.}
    \item{\code{FMCE7}}{A numeric value of expert rating to the FMCE item number 7.}
    \item{\code{FMCE8}}{A numeric value of expert rating to the FMCE item number 8.}
    \item{\code{FMCE9}}{A numeric value of expert rating to the FMCE item number 9.}
    \item{\code{FMCE10}}{A numeric value of expert rating to the FMCE item number 10.}
    \item{\code{FMCE11}}{A numeric value of expert rating to the FMCE item number 11.}
    \item{\code{FMCE12}}{A numeric value of expert rating to the FMCE item number 12.}
    \item{\code{FMCE13}}{A numeric value of expert rating to the FMCE item number 13.}
    \item{\code{FMCE14}}{A numeric value of expert rating to the FMCE item number 14.}
    \item{\code{FMCE15}}{A numeric value of expert rating to the FMCE item number 15.}
    \item{\code{FMCE16}}{A numeric value of expert rating to the FMCE item number 16.}
    \item{\code{FMCE17}}{A numeric value of expert rating to the FMCE item number 17.}
    \item{\code{FMCE18}}{A numeric value of expert rating to the FMCE item number 18.}
    \item{\code{FMCE19}}{A numeric value of expert rating to the FMCE item number 19.}
    \item{\code{FMCE20}}{A numeric value of expert rating to the FMCE item number 20.}
    \item{\code{FMCE21}}{A numeric value of expert rating to the FMCE item number 21.}
    \item{\code{FMCE22}}{A numeric value of expert rating to the FMCE item number 22.}
    \item{\code{FMCE23}}{A numeric value of expert rating to the FMCE item number 23.}
    \item{\code{FMCE24}}{A numeric value of expert rating to the FMCE item number 24.}
    \item{\code{FMCE25}}{A numeric value of expert rating to the FMCE item number 25.}
    \item{\code{FMCE26}}{A numeric value of expert rating to the FMCE item number 26.}
    \item{\code{FMCE27}}{A numeric value of expert rating to the FMCE item number 27.}
    \item{\code{FMCE28}}{A numeric value of expert rating to the FMCE item number 28.}
    \item{\code{FMCE29}}{A numeric value of expert rating to the FMCE item number 29.}
    \item{\code{FMCE30}}{A numeric value of expert rating to the FMCE item number 30.}
    \item{\code{FMCE31}}{A numeric value of expert rating to the FMCE item number 31.}
    \item{\code{FMCE32}}{A numeric value of expert rating to the FMCE item number 32.}
    \item{\code{FMCE33}}{A numeric value of expert rating to the FMCE item number 33.}
    \item{\code{FMCE34}}{A numeric value of expert rating to the FMCE item number 34.}
    \item{\code{FMCE35}}{A numeric value of expert rating to the FMCE item number 35.}
    \item{\code{FMCE36}}{A numeric value of expert rating to the FMCE item number 36.}
    \item{\code{FMCE37}}{A numeric value of expert rating to the FMCE item number 37.}
    \item{\code{FMCE38}}{A numeric value of expert rating to the FMCE item number 38.}
    \item{\code{FMCE39}}{A numeric value of expert rating to the FMCE item number 39.}
    \item{\code{FMCE40}}{A numeric value of expert rating to the FMCE item number 40.}
    \item{\code{FMCE41}}{A numeric value of expert rating to the FMCE item number 41.}
    \item{\code{FMCE42}}{A numeric value of expert rating to the FMCE item number 42.}
    \item{\code{FMCE43}}{A numeric value of expert rating to the FMCE item number 43.}
    \item{\code{FMCE44}}{A numeric value of expert rating to the FMCE item number 44.}
    \item{\code{FMCE45}}{A numeric value of expert rating to the FMCE item number 45.}
    \item{\code{FMCE46}}{A numeric value of expert rating to the FMCE item number 46.}
    \item{\code{FMCE47}}{A numeric value of expert rating to the FMCE item number 47.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("FMCEcontentvalidity")
# Open the data
FMCEcontentvalidity
}
