\name{FMCIcontentvalidity}
\alias{FMCIcontentvalidity}
\docType{data}
\title{The Fluid Mechanics Concept Inventory (FMCI) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the FMCI.
}
\usage{data(FMCIcontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 30 FMCI items. The FMCI numbering system starts from the item 3.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{FMCI3}}{A numeric value of expert rating to the FMCI item number 3.}
    \item{\code{FMCI4}}{A numeric value of expert rating to the FMCI item number 4.}
    \item{\code{FMCI5}}{A numeric value of expert rating to the FMCI item number 5.}
    \item{\code{FMCI6}}{A numeric value of expert rating to the FMCI item number 6.}
    \item{\code{FMCI7}}{A numeric value of expert rating to the FMCI item number 7.}
    \item{\code{FMCI8}}{A numeric value of expert rating to the FMCI item number 8.}
    \item{\code{FMCI9}}{A numeric value of expert rating to the FMCI item number 9.}
    \item{\code{FMCI10}}{A numeric value of expert rating to the FMCI item number 10.}
    \item{\code{FMCI11}}{A numeric value of expert rating to the FMCI item number 11.}
    \item{\code{FMCI12}}{A numeric value of expert rating to the FMCI item number 12.}
    \item{\code{FMCI13}}{A numeric value of expert rating to the FMCI item number 13.}
    \item{\code{FMCI14}}{A numeric value of expert rating to the FMCI item number 14.}
    \item{\code{FMCI15}}{A numeric value of expert rating to the FMCI item number 15.}
    \item{\code{FMCI16}}{A numeric value of expert rating to the FMCI item number 16.}
    \item{\code{FMCI17}}{A numeric value of expert rating to the FMCI item number 17.}
    \item{\code{FMCI18}}{A numeric value of expert rating to the FMCI item number 18.}
    \item{\code{FMCI19}}{A numeric value of expert rating to the FMCI item number 19.}
    \item{\code{FMCI20}}{A numeric value of expert rating to the FMCI item number 20.}
    \item{\code{FMCI21}}{A numeric value of expert rating to the FMCI item number 21.}
    \item{\code{FMCI22}}{A numeric value of expert rating to the FMCI item number 22.}
    \item{\code{FMCI23}}{A numeric value of expert rating to the FMCI item number 23.}
    \item{\code{FMCI24}}{A numeric value of expert rating to the FMCI item number 24.}
    \item{\code{FMCI25}}{A numeric value of expert rating to the FMCI item number 25.}
    \item{\code{FMCI26}}{A numeric value of expert rating to the FMCI item number 26.}
    \item{\code{FMCI27}}{A numeric value of expert rating to the FMCI item number 27.}
    \item{\code{FMCI28}}{A numeric value of expert rating to the FMCI item number 28.}
    \item{\code{FMCI29}}{A numeric value of expert rating to the FMCI item number 29.}
    \item{\code{FMCI30}}{A numeric value of expert rating to the FMCI item number 30.}
    \item{\code{FMCI31}}{A numeric value of expert rating to the FMCI item number 31.}
    \item{\code{FMCI32}}{A numeric value of expert rating to the FMCI item number 32.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Martin, J., Mitchell, J. & Newell, T. Development of a concept inventory for fluid mechanics. in \emph{Proceedings of the 33rd Annual Frontiers in Education 2003 vol. 1 T3D} (IEEE, 2003).
}
\examples{
data(FMCIcontentvalidity)
}
\keyword{datasets}
