\name{FMCIkey}
\alias{FMCIkey}
\docType{data}
\title{The Fluid Mechanics Concept Inventory (FMCI) key dataset}
\description{
The answers key to analyze the students' obtained score on the FMCI.
}
\usage{data(FMCIkey)}
\format{
  A data frame of a key record on the following 30 FMCI items. The FMCI numbering system starts from the item 3.
  \describe{
    \item{\code{FMCI3}}{A key of the FMCI item number 3.}
    \item{\code{FMCI4}}{A key of the FMCI item number 4.}
    \item{\code{FMCI5}}{A key of the FMCI item number 5.}
    \item{\code{FMCI6}}{A key of the FMCI item number 6.}
    \item{\code{FMCI7}}{A key of the FMCI item number 7.}
    \item{\code{FMCI8}}{A key of the FMCI item number 8.}
    \item{\code{FMCI9}}{A key of the FMCI item number 9.}
    \item{\code{FMCI10}}{A key of the FMCI item number 10.}
    \item{\code{FMCI11}}{A key of the FMCI item number 11.}
    \item{\code{FMCI12}}{A key of the FMCI item number 12.}
    \item{\code{FMCI13}}{A key of the FMCI item number 13.}
    \item{\code{FMCI14}}{A key of the FMCI item number 14.}
    \item{\code{FMCI15}}{A key of the FMCI item number 15.}
    \item{\code{FMCI16}}{A key of the FMCI item number 16.}
    \item{\code{FMCI17}}{A key of the FMCI item number 17.}
    \item{\code{FMCI18}}{A key of the FMCI item number 18.}
    \item{\code{FMCI19}}{A key of the FMCI item number 19.}
    \item{\code{FMCI20}}{A key of the FMCI item number 20.}
    \item{\code{FMCI21}}{A key of the FMCI item number 21.}
    \item{\code{FMCI22}}{A key of the FMCI item number 22.}
    \item{\code{FMCI23}}{A key of the FMCI item number 23.}
    \item{\code{FMCI24}}{A key of the FMCI item number 24.}
    \item{\code{FMCI25}}{A key of the FMCI item number 25.}
    \item{\code{FMCI26}}{A key of the FMCI item number 26.}
    \item{\code{FMCI27}}{A key of the FMCI item number 27.}
    \item{\code{FMCI28}}{A key of the FMCI item number 28.}
    \item{\code{FMCI29}}{A key of the FMCI item number 29.}
    \item{\code{FMCI30}}{A key of the FMCI item number 30.}
    \item{\code{FMCI31}}{A key of the FMCI item number 31.}
    \item{\code{FMCI32}}{A key of the FMCI item number 32.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("FMCIkey")
# Open the data
FMCIkey
}
