\name{literacy}
\alias{literacy}
\docType{data}
\title{Students' literacy dataset}
\description{
In this package, students' literacy was defined based on two close-ended items asking the accessibility to available books in their home and digital facilities belong to the students.
}
\usage{data(literacy)}
\format{
  A data frame with 497 observations on the following 2 literacy items.
  \describe{
    \item{\code{LIT1}}{Do you read books in your home? 1 = Yes, 2 = No}
    \item{\code{LIT2}}{Do you access gadgets and the internet in your home? 1 = Yes, 2 = No}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
%\references{
%
%}
\examples{
data(literacy)
}
\keyword{datasets}
