\name{MWCScontentvalidity}
\alias{MWCScontentvalidity}
\docType{data}
\title{The Mechanical Waves Conceptual Survey (MWCS) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the MWCS.
}
\usage{data(MWCScontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 22 MWCS items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{MWCS1}}{A numeric value of expert rating to the MWCS item number 1.}
    \item{\code{MWCS2}}{A numeric value of expert rating to the MWCS item number 2.}
    \item{\code{MWCS3}}{A numeric value of expert rating to the MWCS item number 3.}
    \item{\code{MWCS4}}{A numeric value of expert rating to the MWCS item number 4.}
    \item{\code{MWCS5}}{A numeric value of expert rating to the MWCS item number 5.}
    \item{\code{MWCS6}}{A numeric value of expert rating to the MWCS item number 6.}
    \item{\code{MWCS7}}{A numeric value of expert rating to the MWCS item number 7.}
    \item{\code{MWCS8}}{A numeric value of expert rating to the MWCS item number 8.}
    \item{\code{MWCS9}}{A numeric value of expert rating to the MWCS item number 9.}
    \item{\code{MWCS10}}{A numeric value of expert rating to the MWCS item number 10.}
    \item{\code{MWCS11}}{A numeric value of expert rating to the MWCS item number 11.}
    \item{\code{MWCS12}}{A numeric value of expert rating to the MWCS item number 12.}
    \item{\code{MWCS13}}{A numeric value of expert rating to the MWCS item number 13.}
    \item{\code{MWCS14}}{A numeric value of expert rating to the MWCS item number 14.}
    \item{\code{MWCS15}}{A numeric value of expert rating to the MWCS item number 15.}
    \item{\code{MWCS16}}{A numeric value of expert rating to the MWCS item number 16.}
    \item{\code{MWCS17}}{A numeric value of expert rating to the MWCS item number 17.}
    \item{\code{MWCS18}}{A numeric value of expert rating to the MWCS item number 18.}
    \item{\code{MWCS19}}{A numeric value of expert rating to the MWCS item number 19.}
    \item{\code{MWCS20}}{A numeric value of expert rating to the MWCS item number 20.}
    \item{\code{MWCS21}}{A numeric value of expert rating to the MWCS item number 21.}
    \item{\code{MWCS22}}{A numeric value of expert rating to the MWCS item number 22.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Barniol, P. & Zavala, G. Mechanical waves conceptual survey: Its modification and conversion to a standard multiple-choice test. \emph{Phys Rev Phys Educ Res 12}, 010107 (2016).
}
\examples{
data(MWCScontentvalidity)
}
\keyword{datasets}
