\name{physicsidentity}
\alias{physicsidentity}
\docType{data}
\title{Students' physics identity dataset}
\description{
In this study, students' physics identity was defined based on two close-ended items asking the students' study time for physics and the experienced family recognition when they are studying physics.
}
\usage{data(physicsidentity)}
\format{
  A data frame with 497 observations on the following 2 physics identity items.
  \describe{
    \item{\code{PHYIDE1}}{When did you study physics outside schools? 1 = Most of the time, 2 = Night before the physics schedule, 3 = Night before exam, 4 = Never}
    \item{\code{PHYIDE2}}{Did your parents support you while studying physics? 1 = Yes, 2 = Lack of parental support, 3 = Extremely no}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Hazari, Z., Sonnert, G., Sadler, P. M. & Shanahan, M. C. Connecting high school physics experiences, outcome expectations, physics identity, and physics career choice: A gender study. \emph{J Res Sci Teach 47}, 978–1003 (2010).
}
\examples{
data(physicsidentity)
}
\keyword{datasets}
