\name{RRMCScontentvalidity}
\alias{RRMCScontentvalidity}
\docType{data}
\title{The Rotational and Rolling Motion Conceptual Survey (RRMCS) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the RRMCS.
}
\usage{data(RRMCScontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 30 RRMCS items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{RRMCS1}}{A numeric value of expert rating to the RRMCS item number 1.}
    \item{\code{RRMCS2}}{A numeric value of expert rating to the RRMCS item number 2.}
    \item{\code{RRMCS3}}{A numeric value of expert rating to the RRMCS item number 3.}
    \item{\code{RRMCS4}}{A numeric value of expert rating to the RRMCS item number 4.}
    \item{\code{RRMCS5}}{A numeric value of expert rating to the RRMCS item number 5.}
    \item{\code{RRMCS6}}{A numeric value of expert rating to the RRMCS item number 6.}
    \item{\code{RRMCS7}}{A numeric value of expert rating to the RRMCS item number 7.}
    \item{\code{RRMCS8}}{A numeric value of expert rating to the RRMCS item number 8.}
    \item{\code{RRMCS9}}{A numeric value of expert rating to the RRMCS item number 9.}
    \item{\code{RRMCS10}}{A numeric value of expert rating to the RRMCS item number 10.}
    \item{\code{RRMCS11}}{A numeric value of expert rating to the RRMCS item number 11.}
    \item{\code{RRMCS12}}{A numeric value of expert rating to the RRMCS item number 12.}
    \item{\code{RRMCS13}}{A numeric value of expert rating to the RRMCS item number 13.}
    \item{\code{RRMCS14}}{A numeric value of expert rating to the RRMCS item number 14.}
    \item{\code{RRMCS15}}{A numeric value of expert rating to the RRMCS item number 15.}
    \item{\code{RRMCS16}}{A numeric value of expert rating to the RRMCS item number 16.}
    \item{\code{RRMCS17}}{A numeric value of expert rating to the RRMCS item number 17.}
    \item{\code{RRMCS18}}{A numeric value of expert rating to the RRMCS item number 18.}
    \item{\code{RRMCS19}}{A numeric value of expert rating to the RRMCS item number 19.}
    \item{\code{RRMCS20}}{A numeric value of expert rating to the RRMCS item number 20.}
    \item{\code{RRMCS21}}{A numeric value of expert rating to the RRMCS item number 21.}
    \item{\code{RRMCS22}}{A numeric value of expert rating to the RRMCS item number 22.}
    \item{\code{RRMCS23}}{A numeric value of expert rating to the RRMCS item number 23.}
    \item{\code{RRMCS24}}{A numeric value of expert rating to the RRMCS item number 24.}
    \item{\code{RRMCS25}}{A numeric value of expert rating to the RRMCS item number 25.}
    \item{\code{RRMCS26}}{A numeric value of expert rating to the RRMCS item number 26.}
    \item{\code{RRMCS27}}{A numeric value of expert rating to the RRMCS item number 27.}
    \item{\code{RRMCS28}}{A numeric value of expert rating to the RRMCS item number 28.}
    \item{\code{RRMCS29}}{A numeric value of expert rating to the RRMCS item number 29.}
    \item{\code{RRMCS30}}{A numeric value of expert rating to the RRMCS item number 30.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("RRMCScontentvalidity")
# Open the data
RRMCScontentvalidity
}
