\name{RRMCSkey}
\alias{RRMCSkey}
\docType{data}
\title{Rotational and Rolling Motion Conceptual Survey (RRMCS) key dataset}
\description{
The answers key to analyze the students' obtained score on the RRMCS.
}
\usage{data(RRMCSkey)}
\format{
  A data frame of a key record on the following 47 RRMCS items.
  \describe{
    \item{\code{RRMCS1}}{A key of the RRMCS item number 1.}
    \item{\code{RRMCS2}}{A key of the RRMCS item number 2.}
    \item{\code{RRMCS3}}{A key of the RRMCS item number 3.}
    \item{\code{RRMCS4}}{A key of the RRMCS item number 4.}
    \item{\code{RRMCS5}}{A key of the RRMCS item number 5.}
    \item{\code{RRMCS6}}{A key of the RRMCS item number 6.}
    \item{\code{RRMCS7}}{A key of the RRMCS item number 7.}
    \item{\code{RRMCS8}}{A key of the RRMCS item number 8.}
    \item{\code{RRMCS9}}{A key of the RRMCS item number 9.}
    \item{\code{RRMCS10}}{A key of the RRMCS item number 10.}
    \item{\code{RRMCS11}}{A key of the RRMCS item number 11.}
    \item{\code{RRMCS12}}{A key of the RRMCS item number 12.}
    \item{\code{RRMCS13}}{A key of the RRMCS item number 13.}
    \item{\code{RRMCS14}}{A key of the RRMCS item number 14.}
    \item{\code{RRMCS15}}{A key of the RRMCS item number 15.}
    \item{\code{RRMCS16}}{A key of the RRMCS item number 16.}
    \item{\code{RRMCS17}}{A key of the RRMCS item number 17.}
    \item{\code{RRMCS18}}{A key of the RRMCS item number 18.}
    \item{\code{RRMCS19}}{A key of the RRMCS item number 19.}
    \item{\code{RRMCS20}}{A key of the RRMCS item number 20.}
    \item{\code{RRMCS21}}{A key of the RRMCS item number 21.}
    \item{\code{RRMCS22}}{A key of the RRMCS item number 22.}
    \item{\code{RRMCS23}}{A key of the RRMCS item number 23.}
    \item{\code{RRMCS24}}{A key of the RRMCS item number 24.}
    \item{\code{RRMCS25}}{A key of the RRMCS item number 25.}
    \item{\code{RRMCS26}}{A key of the RRMCS item number 26.}
    \item{\code{RRMCS27}}{A key of the RRMCS item number 27.}
    \item{\code{RRMCS28}}{A key of the RRMCS item number 28.}
    \item{\code{RRMCS29}}{A key of the RRMCS item number 29.}
    \item{\code{RRMCS30}}{A key of the RRMCS item number 30.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). Mendeley Data, V1, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Rimoldini, L. G. & Singh, C. Student understanding of rotational and rolling motion concepts. Physical Review Special Topics - Physics Education Research 1, 010102 (2005).
}
\examples{
data(RRMCSkey)
}
\keyword{datasets}
