\name{STPFASL}
\alias{STPFASL}
\docType{data}
\title{The Survey of Thermodynamic Processes and First and Second Laws (STPFASL) dataset}
\description{
The STPFASL instrument demonstrates 33 items in multiple-choice format that are written based on common student difficulties of thermodynamics as resources in that the incorrect answers to the multiple-choice questions were guided by them.
}
\usage{data(STPFASL)}
\format{
  A data frame of students' responses on the following 47 STPFASL multiple choice items.
  \describe{
    \item{\code{STPFASL1}}{A student's response on the STPFASL item number 1.}
    \item{\code{STPFASL2}}{A student's response on the STPFASL item number 2.}
    \item{\code{STPFASL3}}{A student's response on the STPFASL item number 3.}
    \item{\code{STPFASL4}}{A student's response on the STPFASL item number 4.}
    \item{\code{STPFASL5}}{A student's response on the STPFASL item number 5.}
    \item{\code{STPFASL6}}{A student's response on the STPFASL item number 6.}
    \item{\code{STPFASL7}}{A student's response on the STPFASL item number 7.}
    \item{\code{STPFASL8}}{A student's response on the STPFASL item number 8.}
    \item{\code{STPFASL9}}{A student's response on the STPFASL item number 9.}
    \item{\code{STPFASL10}}{A student's response on the STPFASL item number 10.}
    \item{\code{STPFASL11}}{A student's response on the STPFASL item number 11.}
    \item{\code{STPFASL12}}{A student's response on the STPFASL item number 12.}
    \item{\code{STPFASL13}}{A student's response on the STPFASL item number 13.}
    \item{\code{STPFASL14}}{A student's response on the STPFASL item number 14.}
    \item{\code{STPFASL15}}{A student's response on the STPFASL item number 15.}
    \item{\code{STPFASL16}}{A student's response on the STPFASL item number 16.}
    \item{\code{STPFASL17}}{A student's response on the STPFASL item number 17.}
    \item{\code{STPFASL18}}{A student's response on the STPFASL item number 18.}
    \item{\code{STPFASL19}}{A student's response on the STPFASL item number 19.}
    \item{\code{STPFASL20}}{A student's response on the STPFASL item number 20.}
    \item{\code{STPFASL21}}{A student's response on the STPFASL item number 21.}
    \item{\code{STPFASL22}}{A student's response on the STPFASL item number 22.}
    \item{\code{STPFASL23}}{A student's response on the STPFASL item number 23.}
    \item{\code{STPFASL24}}{A student's response on the STPFASL item number 24.}
    \item{\code{STPFASL25}}{A student's response on the STPFASL item number 25.}
    \item{\code{STPFASL26}}{A student's response on the STPFASL item number 26.}
    \item{\code{STPFASL27}}{A student's response on the STPFASL item number 27.}
    \item{\code{STPFASL28}}{A student's response on the STPFASL item number 28.}
    \item{\code{STPFASL29}}{A student's response on the STPFASL item number 29.}
    \item{\code{STPFASL30}}{A student's response on the STPFASL item number 30.}
    \item{\code{STPFASL31}}{A student's response on the STPFASL item number 31.}
    \item{\code{STPFASL32}}{A student's response on the STPFASL item number 32.}
    \item{\code{STPFASL33}}{A student's response on the STPFASL item number 33.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Brown, B. & Singh, C. Development and validation of a conceptual survey instrument to evaluate students’ understanding of thermodynamics. \emph{Phys Rev Phys Educ Res 17}, 010104 (2021).
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("STPFASL")
# Open the data
STPFASL
}
