\name{STPFASLcontentvalidity}
\alias{STPFASLcontentvalidity}
\docType{data}
\title{The Survey of Thermodynamic Processes and First and Second Laws (STPFASL) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the STPFASL.
}
\usage{data(STPFASLcontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 33 STPFASL items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{STPFASL1}}{A numeric value of expert rating to the STPFASL item number 1.}
    \item{\code{STPFASL2}}{A numeric value of expert rating to the STPFASL item number 2.}
    \item{\code{STPFASL3}}{A numeric value of expert rating to the STPFASL item number 3.}
    \item{\code{STPFASL4}}{A numeric value of expert rating to the STPFASL item number 4.}
    \item{\code{STPFASL5}}{A numeric value of expert rating to the STPFASL item number 5.}
    \item{\code{STPFASL6}}{A numeric value of expert rating to the STPFASL item number 6.}
    \item{\code{STPFASL7}}{A numeric value of expert rating to the STPFASL item number 7.}
    \item{\code{STPFASL8}}{A numeric value of expert rating to the STPFASL item number 8.}
    \item{\code{STPFASL9}}{A numeric value of expert rating to the STPFASL item number 9.}
    \item{\code{STPFASL10}}{A numeric value of expert rating to the STPFASL item number 10.}
    \item{\code{STPFASL11}}{A numeric value of expert rating to the STPFASL item number 11.}
    \item{\code{STPFASL12}}{A numeric value of expert rating to the STPFASL item number 12.}
    \item{\code{STPFASL13}}{A numeric value of expert rating to the STPFASL item number 13.}
    \item{\code{STPFASL14}}{A numeric value of expert rating to the STPFASL item number 14.}
    \item{\code{STPFASL15}}{A numeric value of expert rating to the STPFASL item number 15.}
    \item{\code{STPFASL16}}{A numeric value of expert rating to the STPFASL item number 16.}
    \item{\code{STPFASL17}}{A numeric value of expert rating to the STPFASL item number 17.}
    \item{\code{STPFASL18}}{A numeric value of expert rating to the STPFASL item number 18.}
    \item{\code{STPFASL19}}{A numeric value of expert rating to the STPFASL item number 19.}
    \item{\code{STPFASL20}}{A numeric value of expert rating to the STPFASL item number 20.}
    \item{\code{STPFASL21}}{A numeric value of expert rating to the STPFASL item number 21.}
    \item{\code{STPFASL22}}{A numeric value of expert rating to the STPFASL item number 22.}
    \item{\code{STPFASL23}}{A numeric value of expert rating to the STPFASL item number 23.}
    \item{\code{STPFASL24}}{A numeric value of expert rating to the STPFASL item number 24.}
    \item{\code{STPFASL25}}{A numeric value of expert rating to the STPFASL item number 25.}
    \item{\code{STPFASL26}}{A numeric value of expert rating to the STPFASL item number 26.}
    \item{\code{STPFASL27}}{A numeric value of expert rating to the STPFASL item number 27.}
    \item{\code{STPFASL28}}{A numeric value of expert rating to the STPFASL item number 28.}
    \item{\code{STPFASL29}}{A numeric value of expert rating to the STPFASL item number 29.}
    \item{\code{STPFASL30}}{A numeric value of expert rating to the STPFASL item number 30.}
    \item{\code{STPFASL31}}{A numeric value of expert rating to the STPFASL item number 31.}
    \item{\code{STPFASL32}}{A numeric value of expert rating to the STPFASL item number 32.}
    \item{\code{STPFASL33}}{A numeric value of expert rating to the STPFASL item number 33.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("STPFASLcontentvalidity")
# Open the data
STPFASLcontentvalidity
}
