\name{STPFASLkey}
\alias{STPFASLkey}
\docType{data}
\title{Survey of Thermodynamic Processes and First and Second Laws (STPFASL) key dataset}
\description{
The answers key to analyze the students' obtained score on the STPFASL.
}
\usage{data(STPFASLkey)}
\format{
  A data frame of a key record on the following 33 STPFASL items.
  \describe{
    \item{\code{STPFASL1}}{A key of the STPFASL item number 1.}
    \item{\code{STPFASL2}}{A key of the STPFASL item number 2.}
    \item{\code{STPFASL3}}{A key of the STPFASL item number 3.}
    \item{\code{STPFASL4}}{A key of the STPFASL item number 4.}
    \item{\code{STPFASL5}}{A key of the STPFASL item number 5.}
    \item{\code{STPFASL6}}{A key of the STPFASL item number 6.}
    \item{\code{STPFASL7}}{A key of the STPFASL item number 7.}
    \item{\code{STPFASL8}}{A key of the STPFASL item number 8.}
    \item{\code{STPFASL9}}{A key of the STPFASL item number 9.}
    \item{\code{STPFASL10}}{A key of the STPFASL item number 10.}
    \item{\code{STPFASL11}}{A key of the STPFASL item number 11.}
    \item{\code{STPFASL12}}{A key of the STPFASL item number 12.}
    \item{\code{STPFASL13}}{A key of the STPFASL item number 13.}
    \item{\code{STPFASL14}}{A key of the STPFASL item number 14.}
    \item{\code{STPFASL15}}{A key of the STPFASL item number 15.}
    \item{\code{STPFASL16}}{A key of the STPFASL item number 16.}
    \item{\code{STPFASL17}}{A key of the STPFASL item number 17.}
    \item{\code{STPFASL18}}{A key of the STPFASL item number 18.}
    \item{\code{STPFASL19}}{A key of the STPFASL item number 19.}
    \item{\code{STPFASL20}}{A key of the STPFASL item number 20.}
    \item{\code{STPFASL21}}{A key of the STPFASL item number 21.}
    \item{\code{STPFASL22}}{A key of the STPFASL item number 22.}
    \item{\code{STPFASL23}}{A key of the STPFASL item number 23.}
    \item{\code{STPFASL24}}{A key of the STPFASL item number 24.}
    \item{\code{STPFASL25}}{A key of the STPFASL item number 25.}
    \item{\code{STPFASL26}}{A key of the STPFASL item number 26.}
    \item{\code{STPFASL27}}{A key of the STPFASL item number 27.}
    \item{\code{STPFASL28}}{A key of the STPFASL item number 28.}
    \item{\code{STPFASL29}}{A key of the STPFASL item number 29.}
    \item{\code{STPFASL30}}{A key of the STPFASL item number 30.}
    \item{\code{STPFASL31}}{A key of the STPFASL item number 31.}
    \item{\code{STPFASL32}}{A key of the STPFASL item number 32.}
    \item{\code{STPFASL33}}{A key of the STPFASL item number 33.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). Mendeley Data, V1, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Brown, B. & Singh, C. Development and validation of a conceptual survey instrument to evaluate students’ understanding of thermodynamics. Phys Rev Phys Educ Res 17, 010104 (2021).
}
\examples{
data(STPFASLkey)
}
\keyword{datasets}
