\name{TCE}
\alias{TCE}
\docType{data}
\title{The Thermal Concept Evaluation (TCE) dataset}
\description{
There are 26 multiple-choice items in the TCE examining students’ alternative concepts on heat, temperature, heat transfer and temperature change, and thermal properties. The alternative concepts are discovered based on former studies, and they are used as a researcher basis to construct the TCE distractors.
}
\usage{data(TCE)}
\format{
  A data frame of students' responses on the following 26 TCE multiple choice items.
  \describe{
    \item{\code{TCE1}}{A student's response on the TCE item number 1.}
    \item{\code{TCE2}}{A student's response on the TCE item number 2.}
    \item{\code{TCE3}}{A student's response on the TCE item number 3.}
    \item{\code{TCE4}}{A student's response on the TCE item number 4.}
    \item{\code{TCE5}}{A student's response on the TCE item number 5.}
    \item{\code{TCE6}}{A student's response on the TCE item number 6.}
    \item{\code{TCE7}}{A student's response on the TCE item number 7.}
    \item{\code{TCE8}}{A student's response on the TCE item number 8.}
    \item{\code{TCE9}}{A student's response on the TCE item number 9.}
    \item{\code{TCE10}}{A student's response on the TCE item number 10.}
    \item{\code{TCE11}}{A student's response on the TCE item number 11.}
    \item{\code{TCE12}}{A student's response on the TCE item number 12.}
    \item{\code{TCE13}}{A student's response on the TCE item number 13.}
    \item{\code{TCE14}}{A student's response on the TCE item number 14.}
    \item{\code{TCE15}}{A student's response on the TCE item number 15.}
    \item{\code{TCE16}}{A student's response on the TCE item number 16.}
    \item{\code{TCE17}}{A student's response on the TCE item number 17.}
    \item{\code{TCE18}}{A student's response on the TCE item number 18.}
    \item{\code{TCE19}}{A student's response on the TCE item number 19.}
    \item{\code{TCE20}}{A student's response on the TCE item number 20.}
    \item{\code{TCE21}}{A student's response on the TCE item number 21.}
    \item{\code{TCE22}}{A student's response on the TCE item number 22.}
    \item{\code{TCE23}}{A student's response on the TCE item number 23.}
    \item{\code{TCE24}}{A student's response on the TCE item number 24.}
    \item{\code{TCE25}}{A student's response on the TCE item number 25.}
    \item{\code{TCE26}}{A student's response on the TCE item number 26.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("TCE")
# Open the data
TCE
}
