\name{TCEcontentvalidity}
\alias{TCEcontentvalidity}
\docType{data}
\title{The Thermal Concept Evaluation (TCE) content validity dataset}
\description{
In this data, five PER experts rated the content validity of the TCE.
}
\usage{data(TCEcontentvalidity)}
\format{
  A data frame of expert ratings on the content validity of 26 TCE items.
  \describe{
    \item{\code{ID}}{An anonymized expert identity.}
    \item{\code{TCE1}}{A numeric value of expert rating to the TCE item number 1.}
    \item{\code{TCE2}}{A numeric value of expert rating to the TCE item number 2.}
    \item{\code{TCE3}}{A numeric value of expert rating to the TCE item number 3.}
    \item{\code{TCE4}}{A numeric value of expert rating to the TCE item number 4.}
    \item{\code{TCE5}}{A numeric value of expert rating to the TCE item number 5.}
    \item{\code{TCE6}}{A numeric value of expert rating to the TCE item number 6.}
    \item{\code{TCE7}}{A numeric value of expert rating to the TCE item number 7.}
    \item{\code{TCE8}}{A numeric value of expert rating to the TCE item number 8.}
    \item{\code{TCE9}}{A numeric value of expert rating to the TCE item number 9.}
    \item{\code{TCE10}}{A numeric value of expert rating to the TCE item number 10.}
    \item{\code{TCE11}}{A numeric value of expert rating to the TCE item number 11.}
    \item{\code{TCE12}}{A numeric value of expert rating to the TCE item number 12.}
    \item{\code{TCE13}}{A numeric value of expert rating to the TCE item number 13.}
    \item{\code{TCE14}}{A numeric value of expert rating to the TCE item number 14.}
    \item{\code{TCE15}}{A numeric value of expert rating to the TCE item number 15.}
    \item{\code{TCE16}}{A numeric value of expert rating to the TCE item number 16.}
    \item{\code{TCE17}}{A numeric value of expert rating to the TCE item number 17.}
    \item{\code{TCE18}}{A numeric value of expert rating to the TCE item number 18.}
    \item{\code{TCE19}}{A numeric value of expert rating to the TCE item number 19.}
    \item{\code{TCE20}}{A numeric value of expert rating to the TCE item number 20.}
    \item{\code{TCE21}}{A numeric value of expert rating to the TCE item number 21.}
    \item{\code{TCE22}}{A numeric value of expert rating to the TCE item number 22.}
    \item{\code{TCE23}}{A numeric value of expert rating to the TCE item number 23.}
    \item{\code{TCE24}}{A numeric value of expert rating to the TCE item number 24.}
    \item{\code{TCE25}}{A numeric value of expert rating to the TCE item number 25.}
    \item{\code{TCE26}}{A numeric value of expert rating to the TCE item number 26.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("TCEcontentvalidity")
# Open the data
TCEcontentvalidity
}
