\name{TCEkey}
\alias{TCEkey}
\docType{data}
\title{The Thermal Concept Evaluation (TCE) key dataset}
\description{
The answers key to analyze the students' obtained score on the TCE.
}
\usage{data(TCEkey)}
\format{
  A data frame of answer keys on the following 26 TCE items.
  \describe{
    \item{\code{TCE1}}{A key of the TCE item number 1.}
    \item{\code{TCE2}}{A key of the TCE item number 2.}
    \item{\code{TCE3}}{A key of the TCE item number 3.}
    \item{\code{TCE4}}{A key of the TCE item number 4.}
    \item{\code{TCE5}}{A key of the TCE item number 5.}
    \item{\code{TCE6}}{A key of the TCE item number 6.}
    \item{\code{TCE7}}{A key of the TCE item number 7.}
    \item{\code{TCE8}}{A key of the TCE item number 8.}
    \item{\code{TCE9}}{A key of the TCE item number 9.}
    \item{\code{TCE10}}{A key of the TCE item number 10.}
    \item{\code{TCE11}}{A key of the TCE item number 11.}
    \item{\code{TCE12}}{A key of the TCE item number 12.}
    \item{\code{TCE13}}{A key of the TCE item number 13.}
    \item{\code{TCE14}}{A key of the TCE item number 14.}
    \item{\code{TCE15}}{A key of the TCE item number 15.}
    \item{\code{TCE16}}{A key of the TCE item number 16.}
    \item{\code{TCE17}}{A key of the TCE item number 17.}
    \item{\code{TCE18}}{A key of the TCE item number 18.}
    \item{\code{TCE19}}{A key of the TCE item number 19.}
    \item{\code{TCE20}}{A key of the TCE item number 20.}
    \item{\code{TCE21}}{A key of the TCE item number 21.}
    \item{\code{TCE22}}{A key of the TCE item number 22.}
    \item{\code{TCE23}}{A key of the TCE item number 23.}
    \item{\code{TCE24}}{A key of the TCE item number 24.}
    \item{\code{TCE25}}{A key of the TCE item number 25.}
    \item{\code{TCE26}}{A key of the TCE item number 26.}
  }
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("TCEkey")
# Open the data
TCEkey
}
