\name{FMCE}
\alias{FMCE}
\docType{data}
\title{Force and Motion Conceptual Evaluation (FMCE) dataset}
\description{
Similar with the FCI, the goal of the FMCE is also intended to measure students’ conceptual understanding of Newtonian mechanics. It comprises of 47 multiple-choice items.
}
\usage{data(FMCE)}
\format{
  A data frame with 497 observations on the following 47 FMCE multiple choice items.
  \describe{
    \item{\code{FMCE1}}{A character of the FMCE item number 1.}
    \item{\code{FMCE2}}{A character of the FMCE item number 2.}
    \item{\code{FMCE3}}{A character of the FMCE item number 3.}
    \item{\code{FMCE4}}{A character of the FMCE item number 4.}
    \item{\code{FMCE5}}{A character of the FMCE item number 5.}
    \item{\code{FMCE6}}{A character of the FMCE item number 6.}
    \item{\code{FMCE7}}{A character of the FMCE item number 7.}
    \item{\code{FMCE8}}{A character of the FMCE item number 8.}
    \item{\code{FMCE9}}{A character of the FMCE item number 9.}
    \item{\code{FMCE10}}{A character of the FMCE item number 10.}
    \item{\code{FMCE11}}{A character of the FMCE item number 11.}
    \item{\code{FMCE12}}{A character of the FMCE item number 12.}
    \item{\code{FMCE13}}{A character of the FMCE item number 13.}
    \item{\code{FMCE14}}{A character of the FMCE item number 14.}
    \item{\code{FMCE15}}{A character of the FMCE item number 15.}
    \item{\code{FMCE16}}{A character of the FMCE item number 16.}
    \item{\code{FMCE17}}{A character of the FMCE item number 17.}
    \item{\code{FMCE18}}{A character of the FMCE item number 18.}
    \item{\code{FMCE19}}{A character of the FMCE item number 19.}
    \item{\code{FMCE20}}{A character of the FMCE item number 20.}
    \item{\code{FMCE21}}{A character of the FMCE item number 21.}
    \item{\code{FMCE22}}{A character of the FMCE item number 22.}
    \item{\code{FMCE23}}{A character of the FMCE item number 23.}
    \item{\code{FMCE24}}{A character of the FMCE item number 24.}
    \item{\code{FMCE25}}{A character of the FMCE item number 25.}
    \item{\code{FMCE26}}{A character of the FMCE item number 26.}
    \item{\code{FMCE27}}{A character of the FMCE item number 27.}
    \item{\code{FMCE28}}{A character of the FMCE item number 28.}
    \item{\code{FMCE29}}{A character of the FMCE item number 29.}
    \item{\code{FMCE30}}{A character of the FMCE item number 30.}
    \item{\code{FMCE31}}{A character of the FMCE item number 31.}
    \item{\code{FMCE32}}{A character of the FMCE item number 32.}
    \item{\code{FMCE33}}{A character of the FMCE item number 33.}
    \item{\code{FMCE34}}{A character of the FMCE item number 34.}
    \item{\code{FMCE35}}{A character of the FMCE item number 35.}
    \item{\code{FMCE36}}{A character of the FMCE item number 36.}
    \item{\code{FMCE37}}{A character of the FMCE item number 37.}
    \item{\code{FMCE38}}{A character of the FMCE item number 38.}
    \item{\code{FMCE39}}{A character of the FMCE item number 39.}
    \item{\code{FMCE40}}{A character of the FMCE item number 40.}
    \item{\code{FMCE41}}{A character of the FMCE item number 41.}
    \item{\code{FMCE42}}{A character of the FMCE item number 42.}
    \item{\code{FMCE43}}{A character of the FMCE item number 43.}
    \item{\code{FMCE44}}{A character of the FMCE item number 44.}
    \item{\code{FMCE45}}{A character of the FMCE item number 45.}
    \item{\code{FMCE46}}{A character of the FMCE item number 46.}
    \item{\code{FMCE47}}{A character of the FMCE item number 47.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). Mendeley Data, V1, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Thornton, R. K. & Sokoloff, D. R. Assessing student learning of Newton’s laws: The Force and Motion Conceptual Evaluation and the Evaluation of Active Learning Laboratory and Lecture Curricula. Am J Phys 66, 338–352 (1998).
}
\examples{
data(FMCE)
}
\keyword{datasets}
