\name{RRMCS}
\alias{RRMCS}
\docType{data}
\title{The Rotational and Rolling Motion Conceptual Survey (RRMCS) dataset}
\description{
The RRMCS could be given to explore students’ ideas in various education levels starting in high school to introductory college. It examines students' understanding of rotational motion and notions associated with it through 30 multiple choice items.
}
\usage{data(RRMCS)}
\format{
  A data frame of students' responses on the following 30 RRMCS multiple choice items.
  \describe{
    \item{\code{RRMCS1}}{A student's response on the RRMCS item number 1.}
    \item{\code{RRMCS2}}{A student's response on the RRMCS item number 2.}
    \item{\code{RRMCS3}}{A student's response on the RRMCS item number 3.}
    \item{\code{RRMCS4}}{A student's response on the RRMCS item number 4.}
    \item{\code{RRMCS5}}{A student's response on the RRMCS item number 5.}
    \item{\code{RRMCS6}}{A student's response on the RRMCS item number 6.}
    \item{\code{RRMCS7}}{A student's response on the RRMCS item number 7.}
    \item{\code{RRMCS8}}{A student's response on the RRMCS item number 8.}
    \item{\code{RRMCS9}}{A student's response on the RRMCS item number 9.}
    \item{\code{RRMCS10}}{A student's response on the RRMCS item number 10.}
    \item{\code{RRMCS11}}{A student's response on the RRMCS item number 11.}
    \item{\code{RRMCS12}}{A student's response on the RRMCS item number 12.}
    \item{\code{RRMCS13}}{A student's response on the RRMCS item number 13.}
    \item{\code{RRMCS14}}{A student's response on the RRMCS item number 14.}
    \item{\code{RRMCS15}}{A student's response on the RRMCS item number 15.}
    \item{\code{RRMCS16}}{A student's response on the RRMCS item number 16.}
    \item{\code{RRMCS17}}{A student's response on the RRMCS item number 17.}
    \item{\code{RRMCS18}}{A student's response on the RRMCS item number 18.}
    \item{\code{RRMCS19}}{A student's response on the RRMCS item number 19.}
    \item{\code{RRMCS20}}{A student's response on the RRMCS item number 20.}
    \item{\code{RRMCS21}}{A student's response on the RRMCS item number 21.}
    \item{\code{RRMCS22}}{A student's response on the RRMCS item number 22.}
    \item{\code{RRMCS23}}{A student's response on the RRMCS item number 23.}
    \item{\code{RRMCS24}}{A student's response on the RRMCS item number 24.}
    \item{\code{RRMCS25}}{A student's response on the RRMCS item number 25.}
    \item{\code{RRMCS26}}{A student's response on the RRMCS item number 26.}
    \item{\code{RRMCS27}}{A student's response on the RRMCS item number 27.}
    \item{\code{RRMCS28}}{A student's response on the RRMCS item number 28.}
    \item{\code{RRMCS29}}{A student's response on the RRMCS item number 29.}
    \item{\code{RRMCS30}}{A student's response on the RRMCS item number 30.}
  }
}
\source{
Santoso, P. H. et al. SPHERE: Students' performance dataset of conceptual understanding, scientific ability, and learning attitude in physics education research (PER). \emph{Mendeley Data, V1}, (2024). doi: 10.17632/88d7m2fv7p.1.
}
\references{
Rimoldini, L. G. & Singh, C. Student understanding of rotational and rolling motion concepts. \emph{Physical Review Special Topics - Physics Education Research 1}, 010102 (2005).
}
\examples{
library(spheredata)
# Call the data from spheredata package
get("RRMCS")
# Open the data
RRMCS
}
